+++ examples/auto_encoder_example.py
@@ -4,8 +4,6 @@
 # Author: Tiankai Yang <tiankaiy@usc.edu>
 # License: BSD 2 clause
 
-from __future__ import division
-from __future__ import print_function
 
 import os
 import sys
