+++ pyod/models/base_dl.py
@@ -4,8 +4,6 @@
 # Author: Tiankai Yang <tiankaiy@usc.edu>
 # License: BSD 2 clause
 
-from __future__ import division
-from __future__ import print_function
 
 import os
 import pickle
@@ -15,7 +13,6 @@ from abc import abstractmethod
 from inspect import isfunction
 
 import numpy as np
-import tensorflow
 import torch
 import tqdm
 from sklearn.utils import check_array
@@ -23,25 +20,6 @@ from sklearn.utils import check_array
 from .base import BaseDetector
 from ..utils.torch_utility import TorchDataset, \
     get_optimizer_by_name, get_criterion_by_name
-
-
-# Old function, deprecat this in the future
-def _get_tensorflow_version():  # pragma: no cover
-    """ Utility function to decide the version of tensorflow, which will
-    affect how to import keras models.
-
-    Returns
-    -------
-    tensorflow version : int
-
-    """
-
-    tf_version = str(tensorflow.__version__)
-    if int(tf_version.split(".")[0]) != 1 and int(
-            tf_version.split(".")[0]) != 2:
-        raise ValueError("tensorflow version error")
-
-    return int(tf_version.split(".")[0]) * 100 + int(tf_version.split(".")[1])
 
 
 class BaseDeepLearningDetector(BaseDetector):
